+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Updated to Spell Engine 1.9 and RPG Inventory 2.10.0
 
+## Additions
+
+- class description field now supports longer descriptions
+
 ## Changes
 
 - removed "class_item_identifier" field from rpg_class JSON files (now always uses the "spell_engine:spell_book" item)